Tidy-up and bug fixing for Top Trumps Builder
Added specialised exceptions for more maintainable code, as opposed to manually assigning messages based on return. Also, adding a bunch of stuff that was missing. In my defense, it was very late when I wrote this crap.
Everything is still untested though. Day #X with no internet access, I believe the walls are watching me and I've run out of bottles to read on the bog.